Rename the core Wrapper extension to Root.
